pyproject.toml
[project]
name = "speech-recognition"
version = "0.1.0"
description = "Speech Recognition"
requires-python = "==3.14.2"
dependencies = [
    "aiogram==3.27.0",
    "aiohttp-socks==0.11.0",
    "filetype==1.2.0",
    "pydantic-settings==2.13.1",
    "python-ffmpeg==2.0.12",
    "vosk==0.3.45",
]

[dependency-groups]
dev = [
    "ruff==0.15.9"
]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "B", "Q", "I"]
ignore = ["B027"]
